[fix] movieTalk 관련 api 수정
